Bug found in dMCdROfRhodMusDetector.  Fixed that and cleaned up code in dMuaDetector with comments. Renamed unit tests for DAW and CAW OneLayerDetectors to include "dMC" in name because added derivative detector tests.
